requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
object-detection==0.1
Pillow==6.1.0
pycparser==2.19
pyOpenSSL==19.0.0
requests==2.22.0
six==1.12.0
slim==0.1
urllib3==1.25.3
